create and style interactive button that opens event creation modal connected to backend, fix  formatting
